requirements.txt
pdfminer.six==20221105
nltk==3.6.5
pandas==1.3.4
reportlab
matplotlib